+++ Pandora/Paradigm/Junction/Composition.hs
@@ -1,6 +1,6 @@
 module Pandora.Paradigm.Junction.Composition (Composition (..)) where
 
 class Composition t where
-	{-# MINIMAL composition #-}
-	type Outline t a :: *
-	composition :: t a -> Outline t a
+	{-# MINIMAL unwrap #-}
+	type Primary t a :: *
+	unwrap :: t a -> Primary t a