+++ Pandora/Paradigm/Junction/Schemes/TUV.hs
@@ -2,7 +2,7 @@ module Pandora.Paradigm.Junction.Schemes.TUV (TUV (..)) where
 
 import Pandora.Core.Functor (Variant (Co, Contra), type (:.:), type (><))
 import Pandora.Core.Morphism ((.), ($))
-import Pandora.Paradigm.Junction.Composition (Composition (Outline, composition))
+import Pandora.Paradigm.Junction.Composition (Composition (Primary, unwrap))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>), (<$$>), (<$$$>), comap))
 import Pandora.Pattern.Functor.Contravariant (Contravariant ((>$<), (>$$<), (>$$$<)), contramap)
 import Pandora.Pattern.Functor.Extractable (Extractable (extract))
@@ -17,8 +17,8 @@ import Pandora.Pattern.Functor.Adjoint (Adjoint (phi, psi))
 newtype TUV ct cu cv t u v a = TUV (t :.: u :.: v >< a)
 
 instance Composition (TUV ct cu cv t u v) where
-	type Outline (TUV ct cu cv t u v) a = t :.: u :.: v >< a
-	composition (TUV x) = x
+	type Primary (TUV ct cu cv t u v) a = t :.: u :.: v >< a
+	unwrap (TUV x) = x
 
 instance (Covariant t, Covariant u, Covariant v)
 	=> Covariant (TUV 'Co 'Co 'Co t u v) where
@@ -58,7 +58,7 @@ instance (Pointable t, Pointable u, Pointable v)
 
 instance (Extractable t, Extractable u, Extractable v)
 	=> Extractable (TUV 'Co 'Co 'Co t u v) where
-	extract = extract . extract . extract . composition
+	extract = extract . extract . extract . unwrap
 
 instance (Avoidable t, Covariant u, Covariant v)
 	=> Avoidable (TUV 'Co 'Co 'Co t u v) where
@@ -78,7 +78,7 @@ instance (Traversable t, Traversable u, Traversable v)
 
 instance (Distributive t, Distributive u, Distributive v)
 	=> Distributive (TUV 'Co 'Co 'Co t u v) where
-	x >>- f = TUV . (distribute <$$>) . (distribute <$>) . distribute $ composition . f <$> x
+	x >>- f = TUV . (distribute <$$>) . (distribute <$>) . distribute $ unwrap . f <$> x
 
 type (:-|:) t u = (Extractable t, Pointable t, Extractable u, Pointable u, Adjoint t u)