+++ Pandora/Paradigm/Junction/Schemes/UT.hs
@@ -2,7 +2,7 @@ module Pandora.Paradigm.Junction.Schemes.UT (UT (..)) where
 
 import Pandora.Core.Functor (Variant (Co), type (:.:), type (><))
 import Pandora.Core.Morphism ((.), ($))
-import Pandora.Paradigm.Junction.Composition (Composition (Outline, composition))
+import Pandora.Paradigm.Junction.Composition (Composition (Primary, unwrap))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>), (<$$>), comap))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Extractable (Extractable (extract))
@@ -21,8 +21,8 @@ import Pandora.Pattern.Object.Monoid (Monoid (zero))
 newtype UT ct cu t u a = UT (u :.: t >< a)
 
 instance Composition (UT ct cu t u) where
-	type Outline (UT ct cu t u) a = u :.: t >< a
-	composition (UT x) = x
+	type Primary (UT ct cu t u) a = u :.: t >< a
+	unwrap (UT x) = x
 
 instance (Covariant t, Covariant u) => Covariant (UT 'Co 'Co t u) where
 	f <$> UT x = UT $ f <$$> x
@@ -31,7 +31,7 @@ instance (Pointable t, Pointable u) => Pointable (UT 'Co 'Co t u) where
 	point = UT . point . point
 
 instance (Extractable t, Extractable u) => Extractable (UT 'Co 'Co t u) where
-	extract = extract . extract . composition
+	extract = extract . extract . unwrap
 
 instance (Covariant t, Avoidable u) => Avoidable (UT 'Co 'Co t u) where
 	empty = UT empty
@@ -52,7 +52,7 @@ instance (Traversable t, Traversable u) => Traversable (UT 'Co 'Co t u) where
 	UT x ->> f = UT <$> x ->>> f
 
 instance (Distributive t, Distributive u) => Distributive (UT 'Co 'Co t u) where
-	x >>- f = UT . comap distribute . distribute $ composition . f <$> x
+	x >>- f = UT . comap distribute . distribute $ unwrap . f <$> x
 
 instance Setoid (u :.: t >< a) => Setoid (UT 'Co 'Co t u a) where
 	UT x == UT y = x == y