+++ Pandora/Paradigm/Junction/Transformer.hs
@@ -5,10 +5,10 @@ import Pandora.Pattern.Functor.Covariant (Covariant)
 import Pandora.Pattern.Functor.Pointable (Pointable)
 
 class Transformer t where
-	{-# MINIMAL lay, equip #-}
-	type Layout (t :: * -> *) (u :: * -> *) = (r :: * -> *) | r -> t u
-	lay :: Covariant u => u ~> Layout t u
-	equip :: Pointable u => t ~> Layout t u
+	{-# MINIMAL lay, wrap #-}
+	type Schema (t :: * -> *) (u :: * -> *) = (r :: * -> *) | r -> t u
+	lay :: Covariant u => u ~> Schema t u
+	wrap :: Pointable u => t ~> Schema t u
 
 infixr 1 :>
-type (:>) t u a = Transformer t => Layout t u a
+type (:>) t u a = Transformer t => Schema t u a